scripts/requirements-remove-bg.txt
rembg>=2.0.50
Pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
